backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
google-cloud-bigquery